GithubAuthenticationConfig: move to kamu_adapter_oauth
